Remove extraneous format key on array objects
This commit removes the "format" key from array objects since any format
annotations on arrays are assumed to be applicable to the array items
rather than to the array itself.  Byte arrays are handled as a special
case since they are represented as byte strings in the CRD YAML rather
than actual arrays.

Closes #141

Signed-off-by: Allain Legacy <allain.legacy@windriver.com>