Fix CRD status generation
The CRD status field doesn't have omitempty on its fields, which means
that they get serialized as null.  However, the fields aren't marked
as optional or nullable in the OpenAPI schema, which means they fail
the client-side OpenAPI validation.  This populates the known ones so
that this doesn't happen.  In the long term, we should figure out a
better solution.