Add optional/required markers
Using +kubebuilder:validation:Optional/Required at the package level sets
the default value for fields in that package to that value, while using the
opposite marker on fields in that package sets them as an exception.
The `+optional` marker is an alias for
`+kubebuilder:validation:Optional`, for interoperability with main
Kubernetes.