Ensure same-line last-comment case is handled
We have special handling to skip end-of-line (same-line) comments while
collecting. However, when such a comment is also the last comment in the
file, we have to take special care, because it'll be associated with a
higher-level AST construct, but we may continue traversing to handle
special cases, so we have to take care to acknowlege that we've reached
the end of the file.