Smarter allOf flattening/validation preservation
This improves our allOf flattening so that it progressively works up the
chain, pushing allOf down until it's no longer needed, or it's at the
most specific point.  Along the way, if two allOfs would be identical,
they'll be merged.

We then make use of this by converting our ref merger to just stick
both halves in an allOf, and then let the allOf flattener deal with it.

The previous behavior of flattening simple allOfs (like the ones to
embedded references) just "fall out" of the new logic.

In simple (non-overlapping) cases, we still end up with the
simplest-looking validation, and in complex cases, we're able to
preserve "structurality" (see the structure schema) KEP by dealing with
those fields explicitly when we need to.  However, just letting the
flattener do its thing often solves the issue.