Avoid data race on package type data
Due to a flawed assumption about dummy packages vs normal packages in
the partial type checker, it was possible to have a data race on the
typecheck data when one package dependended on a given package, and
another had its dependency on that package filtered out.

Makes sure to lock when populating the dummy package data to avoid this.