Avoid processing several time the same groupKind
FindKubeKinds may return duplicates if there are several versions of the
same groupKind for which we want to generate CRDs.
When combined with the trvialVersions option, which does mutate a
pointer of the corresponding object, we end up overriding the
top-level `validation` field with `nil`, resulting from the already
overridden schema of another version of the same GroupKind.

This commit fixes it by making sure FindKubeKinds does not return
duplicate entries.