Support non-hashable Generators in genall
It was previous possible for a generator implementation to become
non-hashable by adding non-hashable fields, and since interface values
compare by deferencing to their underlying values, this was a problem.

This changes it to refer to everything by a *pointer* to a Generator,
which just compares the pointer value.  Since we're shipping around
unique instances anyway, this is fine.