preserveUnknownFields: true on v1beta1, option
To avoid a breaking change in crd-gen, we need to default to
`preserveUnknownFields: true` if you're just generating v1beta1 CRDs.

In all other cases (multiple versions, v1 only), we default to
`preserveUnknownFields: false`, as this is the reccomended default and
required by v1.

We also introduce an option to crd-gen to allow overriding this
(`crd:preserveUnknownFields`).