✨ Generate schema for types with custom JSON marshaling as Any type
Currently controller-gen complains about structs with missing json tags
even when those structs implement custom JSON marshalling.

With this change we check if a type implements custom JSON marshalling and if it
does, we output schema for Any type. This still allows the validation type to
be overriden with a marker.