loader.TypeChecker: Don't allow setting a NodeFilter after initialization
Changing the filter invalidates the "has this package already been
checked" list; it will absolutely not do the right thing, so don't
allow changing the filter.

Now, because the different generators each get their own
GenerationContext, nothing currently actually changes the TypeChecker.
But it's a footgun waiting to go off.