Take union of all filters when typechecking
This takes the union of all node filters when type-checking, so that
different generators can't interfere with each other.  Previously, if
you ran a generator with a narrower type filter before one with a broad
filter, you'd get weird results because the narrower filter is the one
that's actually applied to all the shared packages.  Now, we gather all
filters ahead of time, and then use their union: if any filter thinks a
reference is important, we consider that reference important for the
whole context.

Concretely, this manifested if you ran the CRD generator before deepcopy
-- the CRD generator has a much narrower node filter, so you'd get
weird deepcopy results.  If you ran them the other way around, it would
be fine.  Now, either way works.