Generate embedded objectmeta in CRDs
make it so embedded ObjectMeta in the CRD get's properly generated if the
generator option generateEmbeddedObjectMeta=true is passed, this is needed because if a
CRD has embedded ObjectMeta in any field and preserveUnknowFields is set to false, all
the metadata will be lost when doing conversion between versions.

By default any embedded ObjectMeta is not generated in the resulting CRD, however the
top level ObjectMeta belonging to the CRD itself is never generated as the kubernetes API
disallows changes to the CRD metadata between conversions.

The generated ObjectMeta is also only a subset of the original set of fields inside
ObjectMeta this is due to the fact that other runtime fields are problematic if they are
being traded with the kubernetes API, such as creationTimeStamp so this only generates
name, namespace, labels, annotations and finalizers which from a design perspective
should be enough.

An example of why this is useful is if a CRD had a volumeClaimTemplate (for an underlying
statefulset) which include ObjectMeta such as Labels, Annotations and/or name which are
meant to be passed to the PVC.

Also addresses the type of FieldsV1 being typed as "Any" and instead types it as "object"