:bug: Allow setting the type for structs that implement json.Marshaler (#601)
* ✨ Generate schema for types with custom JSON marshaling as Any type

Currently controller-gen complains about structs with missing json tags
even when those structs implement custom JSON marshalling.

With this change we check if a type implements custom JSON marshalling and if it
does, we output schema for Any type. This still allows the validation type to
be overriden with a marker.

* :running: Regenerate pkg/crd/testdata testdata.kubebuilder.io_cronjobs.yaml

* :bug: Allow overriding type of types with custom unmarshaler

This slightly extends the initial work in PR 427 to fall back to the
standard generation rather than producing a schema with Any in it, as
that is invalid and led us to revert that PR.

Co-authored-by: Slavomir Kaslev <kaslevs@vmware.com>