fix: Don't flatten schemas for type idents we don't know about
I hit this problem with cases where there are multiple packages with the
same group and version. The schema patcher would hit a nil pointer
trying to flatten a schema for a kind that exists in one package, but
not another, because it seems to assume that everything's in one
package. By skipping cases where the parser hasn't already found the
given kind in the given package, we got past that panic and everything
works correctly.

I believe this should address #624, though I can't be sure if this is
the only path that can lead to that panic.

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>