Update marker.Collector to fix bug of markers
Markers are saved per `*ast.TypeSpec`, which is specific to `Package` object.
`*ast.TypeSpec` in another `Package` object is different even when
the `Package.ID` is the same, and so collected markers cannot be reused
for another `Package` object.

Therefore, `Package` itself is used as a key for cache of markers
insetead of `Package.ID`.

For more details about the bug, see #783.