:sparkles: testdata/cronjob_types.go: add NestedStructWithSeveralFieldsDoubleMarked (#704)
* testdata/cronjob_types.go: drop field annotation of NestedStructWithSeveralFields

The type `NestedStructWithSeveralFields` is already annotated.
Annotating the field of the same name is not necessary, might confuse
user looking at this example and potentially cover up future regressions
(the field annotation might cover up the type annotation not having the
wanted effect).

Signed-off-by: Jan Fajerski <jfajersk@redhat.com>

* testdata/cronjob_types.go: add NestedStructWithSeveralFieldsDoubleMarked

This adds a field to the cronjob test data, that ensures that fields
which are marked twice (once through a type marker, one through a field
marker) still generate the correct yaml structure.

Signed-off-by: Jan Fajerski <jfajersk@redhat.com>

---------

Signed-off-by: Jan Fajerski <jfajersk@redhat.com>