:warning: Add support for encoding.TextMarshaler (#1015)
* :warning: Add support for encoding.TextMarshaler

Whenever a type is encountered that implements encoding.TextMarshaler
but not encoding/json.Marshaler, assume that it will be encoded as
a string.

This is a breaking change, as types that implement TextMarshaler are now
handled differently. On the other hand, this probably restores reality
for all but the most customized setups, as Go's JSON package will also
check for the presence of this interface, and output a JSON string.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>

* :seedling: More tests for encoding.TextMarshaler

Test that json.Marshaler has precedence over encoding.TextMarshaler.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>

---------

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>