pkg/crd: fix alias type parsing
Using Underlying() was working for basic types but was resolving "too
far" for more complex type, as for the structure. I.e. instead of
retrieving the new type for aliased, it actually retrieved the
underyling (who could have guessed) actual type.

For example, for a string alias, we have:
- Alias type info:           testdata.kubebuilder.io/cronjob.StringAlias
- Rhs from the alias def:    string
- Underlying from the alias: string

But for a struct alias, we have:
- Alias type info:           testdata.kubebuilder.io/cronjob.InlineAlias
- Rhs from the alias def:    testdata.kubebuilder.io/cronjob.EmbeddedStruct
- Underlying from the alias: struct{FromEmbedded string "json:\"fromEmbedded,omitempty\""}

I also imagine that we don't handle nested alias, but this is for
another patch.

Signed-off-by: Mahe Tardy <mahe.tardy@gmail.com>